move ak45 test-only helpers out of driver source into test helpers file
