+++ tests/unit/ak45_driver_test_helpers.h
@@ -1,0 +1,9 @@
+#ifndef AK45_DRIVER_TEST_HELPERS
+#define AK45_DRIVER_TEST_HELPERS
+
+#include <stdint.h>
+
+void ak45_test_reset(void);
+void ak45_set_tx_errors(uint32_t count);
+
+#endif // AK45_DRIVER_TEST_HELPERS
